fix: Don't clobber gopath except for the target that uses the var (#11)
It seems to me that setting gopath would be undesirable since most people should have it set from setting up golang. This just moves those vars into go-compile so that it doesn't bother those who are just building a binary. It might be better to have it only set GOPATH if unset, or move it to .bashrc and friends.